change build to one file
