+++ src/chrome-plugin/content.js
@@ -1,0 +1,5 @@
+import {initCsfd} from "../csfd";
+import {initSeznam} from "../seznam";
+
+initCsfd();
+initSeznam();
