--- src/chrome-plugin/seznam.js
@@ -1,9 +1,0 @@
-import {improvePageResult} from "../seznam/search";
-import {State} from "../seznam/state";
-
-console.debug('---------------- KINOBOX-SEZNAM ----------------')
-
-const state = new State();
-improvePageResult(state);
-
-console.debug('---------------- /KINOBOX-SEZNAM ----------------')
